Mars Creation
Created Mars. Shaped out the the Gym a little more to accurately allow to test the movement, jump height, and jump distance. On Mars, created a little bit of a course in order to test. Added shortcut to obj_Game in order to have us quick jump to Mars. Camera on Mars didn't seem to track the obj_Player_Pilot though. Not sure why. Getting Headache, need to do something different.