Added Quest Tab Logic and Ship Fixes
Added logic to obj_game to get the quest menu working. At the moment, it doesn't do anything except print a debugging line. Added mechanism to flip the ship so we aren't flying upside down, added the background to the space area, and then did some minor testing on the jump to see if I could figure out why the jump was getting locked in place.