Hooked up Hazard object, game over logic
Hooked up the checkpoint and hazard logic. If player collides with the hazard, it will reset to the checkpoint. Also hooked up the hp bar to the hazard and created a game over screen if you run out of health. Need to double check some testing, as the obj_Game object isn't destroyed, so it is *possible* that it might double up if the player restarts.