Made Text Boxes and Potentially fixed Collision
Made text boxes. Used in room, the text collider will create and edit the textbox to the field we ask it to.

Also might have fixed collision. Changed collider to match the player sprite and not to update with animation so now when it turns the player's collider doesn't change. I think this fixed it, but not 100% sure. Please test.