Audio Implemented
Added Sound effect (basic) for the laser weapon. Additionally, added audio controller and a collider that will change audio on collision to a target track. Added two samples for testing. Some minor fading issues, but it seems to not fully "stop" the track if it is the same track? Unsure. Might need to play around with it more.